Haven't done this yet:(Need to get a sales breakdown overview. Also, need custom form inputting (for rands or potentially dollars in the future) for models not using the model form. Need to add a form wrapper per field so users can add other models whilst adding models. Need to reposition the header logout button position.)
Need menu widget in the header. All models need date added, and need to add quote number for MPT to sale model.

Added sale probability and status. Added a manager overview of a person's sales, as well as the ability to edit those sales. Required to add editing of meeting, company, company representative, and to neaten by adding files to group all this crap together. Buttons also need to be made to loo smaller, and need to neaten the looks of the files. Finally, delete sale, meeting, company and company representative needs to be added.

Signed-off-by: Matthew Bezuidenhout <mattpetebez@gmail.com>