Haven't done this yet:
 Need to get a sales breakdown overview. Also, need custom form inputting (for rands or potentially dollars in the future) for models not using the model form. Need to add a form wrapper per field so users can add other models whilst adding models.

Need menu widget in the header. All models need date added.

Required to add editing of company, company representative, and to neaten by adding files to group all this crap together. Finally, delete sale, meeting, company and company representative needs to be added.

Required to edit the success edit and added pages for all models to allow a link back to the profile.

Moved the header links to the right, and added a "View My Sales" and "My Dash Board" link to the header bar.

Signed-off-by: Matthew Bezuidenhout <mattpetebez@gmail.com>