Haven't done this yet:
 Need to get a sales breakdown overview. Also, need custom form inputting (for rands or potentially dollars in the future) for models not using the model form. Need to add a form wrapper per field so users can add other models whilst adding models.

Need menu widget in the header.

Required to add editing of company, company representative, and to neaten by adding files to group all this crap together. Delete and view company, company rep and others is required.

Still need to fix the delete return html, or develop my own one.

Done:

Added date added to all models.

Added model Goals, SalesTeam, and removed SaleProbability and SaleStatus in favour of choices text fields in Sale. Added nullable fields for everything (so as to manage later and prevent fixing the data before setting the schema). Changed meetings to activities, and gave it choices for its purpose and type. Added a consultant company and consultant to sales, as this was a requirement.

Need to add the view for admins, team leaders and team memebers for the index (to their dashboard).

Signed-off-by: Matthew Bezuidenhout <mattpetebez@gmail.com>